move bitmap to bitarray package
