+++ .gitignore
@@ -1,2 +1,3 @@
 *.out
 *.test
+.idea
